pyproject.toml
[project]
name = "mdm-gcp"
version = "0.2.0"
description = "Master Data Management on Google Cloud Platform - Batch & Streaming Processing Pipeline"
authors = [
    {name = "johanesalxd", email = "17249308+alexanderv21@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    # Core data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=12.0.0",

    # Google Cloud Platform - BigQuery
    "google-cloud-bigquery>=3.11.0",
    "google-auth>=2.22.0",
    "db-dtypes>=1.1.0",

    # Google Cloud Platform - Spanner (for streaming)
    "google-cloud-spanner>=3.40.0",

    # Data visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",

    # Jupyter and notebook support
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "ipywidgets>=8.0.0",

    # Utilities
    "faker>=19.0.0",
    "psutil>=5.9.0",
    "tqdm>=4.62.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["batch_mdm_gcp", "streaming_mdm_gcp"]
